requirements.txt
boto3==1.18.25
botocore==1.21.25
jmespath==0.10.0
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
